+++ backend/src/test/java/de/metanome/backend/input/database/ResultSetFixture.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package de.metanome.backend.input.sql;
+package de.metanome.backend.input.database;
 
 import com.mysql.jdbc.ResultSetMetaData;
