+++ backend/src/test/java/de/metanome/backend/input/database/ResultSetTwoLinesFixture.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package de.metanome.backend.input.sql;
+package de.metanome.backend.input.database;
 
 import com.google.common.collect.ImmutableList;
