+++ backend/src/test/java/de/metanome/backend/input/file/CsvFileShortLineFixture.java
@@ -14,26 +14,24 @@
  * limitations under the License.
  */
 
-package de.metanome.backend.input.csv;
+package de.metanome.backend.input.file;
 
 import com.google.common.base.Joiner;
+import com.google.common.collect.ImmutableList;
 
 import de.metanome.algorithm_integration.configuration.ConfigurationSettingFileInput;
 import de.metanome.algorithm_integration.input.InputGenerationException;
 import de.metanome.algorithm_integration.input.InputIterationException;
 
 import java.io.StringReader;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 /**
- * A fixture generating a csv file with 4 rows. Rows 2 and 4 have differing lengths (2 (short) and 4
+ * A fixture generating a file file with 4 rows. Rows 2 and 4 have differing lengths (2 (short) and 4
  * (long)).
  *
- * @author Tanja Bergmann
+ * @author Jakob Zwiener
  */
-public class CsvFileNullValuesFixture {
+public class CsvFileShortLineFixture {
 
   protected static final char QUOTE_CHAR = '\'';
   protected static final char SEPARATOR = ',';
@@ -60,39 +58,20 @@ public class CsvFileNullValuesFixture {
         .setSkipDifferingLines(skipDifferingLines);
 
     return new FileIterator("some_file",
-                            new StringReader(Joiner.on(',').join(getFirstLineWithEmptyStrings()) + "\n" +
-                         Joiner.on(',').join(getSecondLineWithEmptyStrings())),
+                            new StringReader(
+                                Joiner.on(',').join(getExpectedFirstParsableLine()) +
+                                "\nfour,five\n" +
+                                Joiner.on(',').join(getExpectedSecondParsableLine()) +
+                                "\nnine,ten,eleven,twelve"),
                             setting);
   }
 
-  public List<String> getFirstLineWithEmptyStrings() {
-    List<String> list = new ArrayList<>();
-    list.add("one");
-    list.add("");
-    list.add("three");
-    return Collections.unmodifiableList(list);
-  }
-  public List<String> getSecondLineWithEmptyStrings() {
-    List<String> list = new ArrayList<>();
-    list.add("four");
-    list.add("five");
-    list.add("");
-    return Collections.unmodifiableList(list);
+  public ImmutableList<String> getExpectedFirstParsableLine() {
+    return ImmutableList.of("one", "two", "three");
   }
 
-  public List<String> getFirstLineWithNullValues() {
-    List<String> list = new ArrayList<>();
-    list.add("one");
-    list.add(null);
-    list.add("three");
-    return Collections.unmodifiableList(list);
-  }
-  public List<String> getSecondLineWithNullValues() {
-    List<String> list = new ArrayList<>();
-    list.add("four");
-    list.add("five");
-    list.add(null);
-    return Collections.unmodifiableList(list);
+  public ImmutableList<String> getExpectedSecondParsableLine() {
+    return ImmutableList.of("six", "seven", "eight");
   }
 
 }
