+++ frontend/src/main/java/de/metanome/frontend/client/datasources/FileInputEditForm.java
@@ -32,7 +32,7 @@ import com.google.gwt.user.client.ui.Widget;
 import au.com.bytecode.opencsv.CSVParser;
 import au.com.bytecode.opencsv.CSVReader;
 
-import de.metanome.backend.input.csv.FileIterator;
+import de.metanome.backend.input.file.FileIterator;
 import de.metanome.backend.results_db.FileInput;
 import de.metanome.frontend.client.TabWrapper;
 import de.metanome.frontend.client.helpers.FilePathHelper;
