+++ testing_algorithms/example_sql_profiling_algorithm/src/main/java/de/metanome/algorithms/testing/example_sql_profiling_algorithm/SqlProfilingAlgorithm.java
@@ -33,7 +33,7 @@ import java.util.ArrayList;
  */
 public class SqlProfilingAlgorithm implements DatabaseConnectionParameterAlgorithm {
 
-  public static final String DATABASE_IDENTIFIER = "sql identifier";
+  public static final String DATABASE_IDENTIFIER = "database identifier";
 
   protected DatabaseConnectionGenerator inputGenerator;
