+++ testing_algorithms/example_sql_profiling_algorithm/src/test/java/de/metanome/algorithms/testing/example_sql_profiling_algorithm/SqlProfilingAlgorithmTest.java
@@ -91,7 +91,7 @@ public class SqlProfilingAlgorithmTest {
                                                           inputGenerator);
     // Expected values
     // TODO add asserts
-    // the systems property should be output from an sql connection
+    // the systems property should be output from an database connection
 
   }
 }
